Move sample images in img folder in public
