+++ frontend/package.json
@@ -11,7 +11,6 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-ios-switch": "^0.1.19",
     "react-router-dom": "^6.23.1",
     "react-scripts": "5.0.1",
     "react-tinder-card": "^1.6.4",
