+++ frontend/src/App.css
@@ -8,7 +8,7 @@
 }
 
 .Home-header {
-  background: url("./images/puppy_home.jpg");
+  background: url("../public/img/puppy_home.jpg");
   background-repeat: no-repeat;
   min-height: 100vh;
   display: flex;
@@ -74,19 +74,19 @@
 
 .card {
   position: relative;
-  /* background-color: #fff; */
+  background-color: transparent;
   width: 80vw;
   max-width: 260px;
   height: 300px;
-  box-shadow: 0px 0px 60px 0px rgba(0, 0, 0, 0.3);
+  box-shadow: 0px 0px 10px 0px rgba(204, 168, 244, 0.1);
   border-radius: 20px;
   background-size: cover;
   background-position: center;
 }
 
 .cardContent {
-  width: 200px;
-  height: 200px;
+  width: 100%;
+  height: 100%;
 }
 
 .swipe:last-of-type {
